mcp-server/pyproject.toml
[project]
name = "invoice-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.4",
    "grpcio>=1.75.0",
    "grpcio-tools>=1.75.0",
    "mcp[cli]>=1.14.1",
]

[tool.uv.workspace]
members = []
